Allow parsing multiple statements
Allows for parsing programs and blocks
